+++ BidmadPluginSample/Assets/Bidmad/Scripts/Bidmad/BidmadCommon.cs
@@ -15,7 +15,7 @@ public enum BidmadTrackingAuthorizationStatus
 
 public class BidmadCommon
 {
-    string UNITY_PLUGIN_VERSION = "2.7.2";
+    string UNITY_PLUGIN_VERSION = "2.8.0";
 #if UNITY_IOS
     [DllImport("__Internal")]
     private static extern void _bidmadSetDebug(bool isDebug);